Styling and Major Process Fix [Lape]

+++ src/process/action.php
@@ -37,11 +37,17 @@ if (isset($_POST['enroll'])) {
     header("Refresh:0; url=./dashboard.php");
 }
 // studentlist.php
+if (isset($_GET['view'])) {
+    $id = $_GET['view'];
+
+    $student = $mysqli->query("SELECT * FROM student_info WHERE Student_ID='$id'") or die("Connection failed:");
+    $student_details = $student->fetch_assoc();
+}
+// studentlist.php
 if (isset($_GET['delete'])) {
     $id = $_GET['delete'];
 
     $mysqli->query("DELETE FROM student_info WHERE Student_ID='$id'") or die("Connection failed:");
-    header("Refresh:0; url=../dashboard/studentlist.php");
 }
 // applicants.php
 if (isset($_GET['approve'])) {
@@ -70,31 +76,27 @@ if (isset($_GET['delete'])) {
 if (isset($_GET['edit'])) {
     $id = $_GET['edit'];
 
-    $mysqli->query("SELECT First_Name, Last_Name, Middle_Initial, Gender, Birthdate, Place_of_Birth, Citizenship, Civil_Status, Mobile_Number, Email, Address FROM student_info WHERE Student_ID = '$id'");
-
-    header("Refresh:0; url=../dashboard/edit_profile.php");
+    $profile = $mysqli->query("SELECT * FROM student_info WHERE Student_ID='$id'") or die("Connection failed:");
+    $profile_details = $profile->fetch_assoc();
 }
 
-//edit profile.php
 if (isset($_POST['update'])) {
 
-    $fname_edit = $_POST['fname'];
-    $mname_edit = $_POST['lname'];
-    $lname_edit = $_POST['mname'];
+    $firstName_edit = $_POST['firstName'];
+    $lastName_edit = $_POST['lastName'];
+    $middleInitial_edit = $_POST['middleInitial'];
     $sex_edit = $_POST['sex'];
-    $bdate_edit = $_POST['birthdate'];
-    $pob_edit = $_POST['birthplace'];
+    $birthdate_edit = $_POST['birthdate'];
+    $placeofBirth_edit = $_POST['birthplace'];
     $citizenship_edit = $_POST['citizenship'];
-    $civilstat_edit = $_POST['civilstatus'];
-    $mobilenum_edit = $_POST['mobilenum'];
+    $civilStatus_edit = $_POST['civilStatus'];
+    $mobileNumber_edit = $_POST['mobileNumber'];
     $email_edit = $_POST['email'];
     $address_edit = $_POST['address'];
-    $enrolled_edit = $_POST['enrolled'];
+    $enrolled_edit = $_POST['enrollmentStatus'];
 
-    $sql = "SELECT * FROM student_info WHERE Student_ID = $row_ID";
-    $result = $mysqli->query($sql);
-
-    $mysqli->query("UPDATE student_info SET First_Name = $fname_edit, Middle_Initial = $mname_edit, Last_Name = $lname_edit, Gender = $sex_edit, Birthdate = $bdate_edit, Place_of_Birth = $pob_edit, Citizenship = $citizenship_edit, Civil_Status = $civilstat_edit, Mobile_Number = $mobilenum_edit, Email = $email_edit, Address = $address_edit, Enrolled = $enrolled_edit WHERE Student_ID = $row_ID") or die("Connection failed:");
-
+    $mysqli->query("UPDATE student_info SET First_Name = '$firstName_edit', Last_Name = '$lastName_edit', Middle_Initial = '$middleInitial_edit', Gender = '$sex_edit', Birthdate = '$birthdate_edit', Place_of_Birth = '$placeofBirth_edit', Citizenship = '$citizenship_edit', Civil_Status = '$civilStatus_edit', Mobile_Number = '$mobileNumber_edit', Email = '$email_edit', Address = '$address_edit', Enrolled = '$enrolled_edit' WHERE Student_ID='$id'") or die("Connection failed:");
     $mysqli->close();
+
+    header("Refresh:0; url=../dashboard/studentlist.php");
 }
